Move link filter to its own package
