+++ pkg/modules/link_filter/link_filter_test.go
@@ -1,11 +1,13 @@
-package modules
+package link_filter
 
 import (
 	"testing"
+
+	"github.com/pajbot/pajbot2/pkg/modules"
 )
 
 func TestLinkFilterUnmatches(t *testing.T) {
-	spec, ok := GetModuleSpec("link_filter")
+	spec, ok := modules.GetModuleSpec("link_filter")
 	if !ok {
 		t.Fatal("AAAAAA")
 	}
